+++ src/biz/bokhorst/xprivacy/ActivityApp.java
@@ -335,11 +335,10 @@ public class ActivityApp extends Activity {
 			public void onClick(DialogInterface dialog, int which) {
 				// Get toggle
 				boolean restricted = false;
-				List<String> listRestriction = PrivacyManager.getRestrictions(true);
+				List<String> listRestriction = PrivacyManager.getRestrictions(false);
 				for (String restrictionName : listRestriction)
 					if (PrivacyManager.getSettingBool(null, ActivityApp.this, 0,
-							String.format("Template.%s", restrictionName),
-							!PrivacyManager.isDangerousRestriction(restrictionName), false))
+							String.format("Template.%s", restrictionName), true, false))
 						if (PrivacyManager.getRestricted(null, ActivityApp.this, mAppInfo.getUid(), restrictionName,
 								null, false, false)) {
 							restricted = true;
@@ -350,8 +349,7 @@ public class ActivityApp extends Activity {
 				restricted = !restricted;
 				for (String restrictionName : listRestriction)
 					if (PrivacyManager.getSettingBool(null, ActivityApp.this, 0,
-							String.format("Template.%s", restrictionName),
-							!PrivacyManager.isDangerousRestriction(restrictionName), false))
+							String.format("Template.%s", restrictionName), true, false))
 						PrivacyManager.setRestricted(null, ActivityApp.this, mAppInfo.getUid(), restrictionName, null,
 								restricted);